refactor data layer - rename …
add ef to data project

+++ PaymentGateway.Models/BankAccount.cs
@@ -5,11 +5,11 @@
         public int Id { get; set; }
         public int PersonId { get; set; }
 
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         public string Currency { get; set; }
         public string Iban { get; set; }
         public string Status { get; set; }
-        public double Limit { get; set; }
+        public decimal Limit { get; set; }
         public string Type { get; set; }
 
     }
